modified google login
