+++ client/src/components/LoginPage/LoginPage.jsx
@@ -1,10 +1,37 @@
-import React , { useState } from 'react';
+import React , { useState , useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import styles from './LoginPage.module.css';
-import Footer from '../Footer/Footer.jsx';
 import { useDispatch } from 'react-redux';
 import { userLogin } from '../../Redux/actions/actionsUserLogin.js';
 import { useNavigate } from 'react-router-dom';
+import Footer from '../Footer/Footer.jsx';
+import styles from './LoginPage.module.css';
+
+//! Autenticación con Google
+
+import GoogleLogin from "react-google-login";
+import { gapi } from "gapi-script";
+
+
+const clientID = "932914293926-uo3dpst96jr8s51di1mmbhdh3j2gie6a.apps.googleusercontent.com";
+   
+const onSuccess = (response) => {
+    console.log(response)
+}
+
+const onFailure = (response) => {
+    console.log("Something went wrong", response)
+}
+// useEffect(() => {
+//     const start = () => {
+//         gapi.auth2.init({
+//             clientId: clientID,
+//         })
+//     }
+//     gapi.load("client:auth2", start);
+// },[])
+
+//! --------------------------------------
+// CSS REACT-BOOSTRAP
 import NavBar from "../NavBar/NavBar.jsx";
 import Form from "react-bootstrap/Form";
 import Row from "react-bootstrap/Row";
@@ -12,7 +39,6 @@ import Col from "react-bootstrap/Col";
 import Button from "react-bootstrap/Button";
 import Title from '../Title/Title';
 import Stack from 'react-bootstrap/esm/Stack';
-// import loginValidate from "./validate/loginValidate.js";
 
 function LoginPage(){
 
@@ -26,7 +52,7 @@ function LoginPage(){
     })
 
 
-        const submitHandler = async (event) => {
+    const submitHandler = async (event) => {
         event.preventDefault();
         const status = await dispatch(userLogin(form.email , form.password ))
         if (status.hasOwnProperty('error')) alert("Incorrect email or password");
@@ -34,15 +60,13 @@ function LoginPage(){
             alert("login successfully");
             navigate("/");
         }
-        //login funcionando
     }
 
     const changeHandler = (event) => {
     const property = event.target.name;
     const value = event.target.value;
-    //! Elimina id delay de la validación
-    setForm({...form, [property]: value});
-}
+        setForm({...form, [property]: value});
+    };
 
 return (
     <div className="container-fluid">
@@ -64,7 +88,6 @@ return (
                         />
                     </Form.Group>
                     <br/>
-
                     <Form.Group as={Col} controlidpassword="Password">
                         <Form.Label>Password</Form.Label>
                         <Form.Control
@@ -80,11 +103,20 @@ return (
                 <br/>
                 <Button variant="success" type="submit">Login</Button>
             </Form>
-            <Link to={"/register"}>Register Now</Link>
-            </div>
+            <Row className="mb-3">
+                <Link to={"/register"}>Register Now</Link>
+            </Row>
+            <Row className="mb-3">
+                <GoogleLogin
+                    clientId={clientID}
+                    onSuccess={onSuccess}
+                    onFailure={onFailure}
+                    cookiePolicy={"single_host_policy"}
+                />
+            </Row>
+        </div> 
         <Footer/>
-    </div>);
-
-}
+    </div>
+)};
 
 export default LoginPage;
